Add Book Info Page
Added the book info page to show the title, description, availability,
and review average. The user can open a book by clicking on the book on
the main page. It then passes the book to the bookinfo class and shows
all the info on the screen. There are currently two dummy buttons: one
for adding a review and the other for requesting the book. There is a
back arrow on the book page to bring the user back to the main page.