Add "Add Book Page" and Search API
Set up the add book page and the search for books with API.
It currently loads the books from the search method into the main page
book list. Eventually will need to add a method for the user to add a
book to the actual book list. Working good so far.