Update Add Book Page
Fixed the add book page to get the isbn for each book and load it into
the list. Then added the APIs necessary to get a book given an ISBN.
This fully works and allows the user to add books to their library.
This commit also allows the user to add books to the library with an
add button that shows up in the book info page if the user is accessing
the info page from either the search for books via title and author,
or the search for a specific book with an ISBN.

This all currently only works in memory. (Adding a book really only
adds the book to the list that is kept in memory). Future updates
will make an API call to add these books to the actual Book Database
AKA Book DB.