Add Barcode Scanner Functionality
I added the barcode scanner functionality that scans book ISBN barcodes.
This essentially does the same thing as entering the ISBN, but it scans
for the ISBN instead of requiring manual entry. Not much more to say about
this feature. It does the exact same thing as manually entering the ISBN.