General Bug Fix
There was a an issue with the way I was routing the pages.
I continually stacked more and more pages onto the stack without
ever removing them, so the camera was constantly going in one page
that was still on the stack. I am now completely resetting the stack
when the user uses the navigation pane or gets a list of books or
a specific book...