Update Entity Classes
I have fully implemented all of the entity classes in the intellij.
This creates the tables for the entire database. Still need to make the
repository and service files to make it fully work properly. After doing
that, we should be able to start handling calls in the controller
section. Will need to implement an auth controller to handle logins and
and token refreshes. Just a little bit to go...