Complete Service and Repository Classes
Fully implemented the service and repository classes for the 9 database
entities. This is ready to start API development now. I also created
2 more controllers: AuthController and PermController...
These do as you would expect. AuthController: handles authentications
PermController: handles changing users, permissions, etc...
Will probably change this name later, but it is a placeholder as is...