Add Login Validation
I added login validation for the api. It now allows basic authorization
with a username and password. It creates a refresh token on valid
credentials and stores it in the database. Need to set up hashing methods
and access tokens for further access. Also need to amend the database
to allow the hash code in the User table so we know what hashing method
to use to check passwords...