Add Form for Adding API Url to Mobile App
Worked on the form for the user to log in to the API through the mobile
app. This is the first step of integrating the two. Still needs some
work. Need to validate the username and password entered by making an
API call to log in. Based on the returned value, we will know whether it
is valid or not.